Add support for loading different MRom binaries
This commit allows to load different multirom binaries from trampoline.
It is enabled by defining MR_BINARY_SELECTOR in a device's Makefile
which points a C source file from the android root directory. In this
file, the function get_multirom_binary_string must be implemented which
writes a string to the passed parameter that is the name of the multirom
binary to load.
Any implementation must follow these conventions:
- On an error, or if the default binary ("multirom") shall be loaded, -1
must be returned.
- The name of the multirom binary may not exceed 255 chars.

The reasoning behind this change is that incompatible changes between
different kernel versions may occur which require multirom binaries to
be compiled for these specific kernel/android versions. Since the main
distribution path via the multirom app does not allow to deliver
different multirom zip files, runtime dispatching is required.

This case has been observed for the Z2 (or similar devices) when
switching from the 1.2.2 to the 1.3.3 AOSP kernel that included a change
in the display driver. This caused a crash in multirom if the binary was
compiled for the other kernel that was not the primary one of the
device.

Change-Id: I76df9e0ac022b5e7d62259e7fac41c3951451ebc
Signed-off-by: Alexander Diewald <Diewi@diewald-net.com>